# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/inquiries": {
            "get": {
                "tags": [
                    "Inquiry"
                ],
                "summary": "Get Inquiry details",
                "description": "Retrieve the information of the inquiry.",
                "operationId": "inquiry",
                "responses": {
                    "200": {
                        "description": "Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inquiries"
                ],
                "summary": "Create a new inquiry",
                "operationId": "3f728b15f306ab8aecbf3a468e2a4a56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inquiry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiry created successfully"
                                        },
                                        "inquiry": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/inquiries/bulk-upload": {
            "post": {
                "tags": [
                    "Inquiry"
                ],
                "summary": "Bulk upload inquiries",
                "description": "Upload a CSV or TXT file to bulk import inquiries into the system.",
                "operationId": "bulkUpload",
                "requestBody": {
                    "description": "CSV or TXT file containing inquiries to upload.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "The CSV or TXT file to upload.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inquiries imported successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inquiries imported successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The file field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during file upload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "There was an error during import."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/international_inquiries": {
            "get": {
                "tags": [
                    "International Inquiry"
                ],
                "summary": "Get International Inquiry details",
                "description": "Retrieve the information of the international inquiry.",
                "operationId": "international_inquiry",
                "responses": {
                    "200": {
                        "description": "International Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "International Inquiries"
                ],
                "summary": "Create a new international inquiry",
                "operationId": "d30ee798018b89defc5f5afc578c250c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "International Inquiry created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "International Inquiry created successfully"
                                        },
                                        "inquiry": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/inquiry-approved-international-offer": {
            "get": {
                "tags": [
                    "International Inquiry"
                ],
                "summary": "Get International Inquiry details",
                "description": "Retrieve the information of the international inquiry.",
                "operationId": "international_approved_inquiry",
                "responses": {
                    "200": {
                        "description": "International Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inquiry-cancellation-international-offers": {
            "get": {
                "tags": [
                    "International Inquiry"
                ],
                "summary": "Get International Inquiry details",
                "description": "Retrieve the information of the international inquiry.",
                "operationId": "international_cancellation_inquiry",
                "responses": {
                    "200": {
                        "description": "International Inquiry details retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}